requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# LangChain and AI
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.3.7

# Email handling
imaplib2==3.6
email-validator==2.1.0

# Database
sqlite3
sqlalchemy==2.0.23

# Audio processing (for call agent)
openai-whisper==20231117
pydub==0.25.1

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Logging and monitoring
structlog==23.2.0

# Enhanced integrations
elevenlabs==0.2.26
langchain-openai==0.0.2
langchain-community==0.0.10

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
